Move b64 to utf8 encoder to util
